fix cute-dsl moe test
