+++ tests/moe/test_cute_dsl_fused_moe.py
@@ -33,18 +33,17 @@ import weakref
 
 import pytest
 import torch
-from torch.nn import functional as F
 
-from flashinfer.tllm_enums import (
-    ActivationType,
-    DEFAULT_SWIGLU_ALPHA,
-    DEFAULT_SWIGLU_BETA,
-    DEFAULT_SWIGLU_LIMIT,
-)
+from flashinfer.tllm_enums import ActivationType
 from flashinfer.fused_moe.cute_dsl.moe_utils import (
     normalize_cute_dsl_moe_activation_type,
 )
 from flashinfer.cute_dsl import is_cute_dsl_available
+from .utils import (
+    check_accuracy,
+    compute_reference_moe_fp4,
+    create_moe_tensors,
+)
 
 
 def is_sm100_family():
@@ -66,392 +65,6 @@ sm100_required = pytest.mark.skipif(
     not is_sm100_family(),
     reason="Requires SM100 family GPU (Blackwell: SM100, SM103, SM110)",
 )
-
-
-def interleave_linear_and_gate(
-    x: torch.Tensor, group_size: int = 64, dim: int = -1
-) -> torch.Tensor:
-    """Interleave linear and gate weights for SwiGLU."""
-    sizes = x.size()
-    dim = dim % x.dim()
-    assert sizes[dim] % (group_size * 2) == 0
-    prev_sizes = sizes[:dim]
-    post_sizes = sizes[dim + 1 :]
-    x = x.view(*prev_sizes, 2, sizes[dim] // (group_size * 2), group_size, *post_sizes)
-    x = x.transpose(dim, dim + 1).contiguous().view(*sizes)
-    return x
-
-
-def quant_dequant_fp4_reference(
-    tensor: torch.Tensor,
-    global_scale: torch.Tensor,
-    sf_vec_size: int = 16,
-) -> torch.Tensor:
-    """Simulate FP4 quantization and dequantization for reference computation."""
-    from flashinfer.fp4_quantization import fp4_quantize, e2m1_and_ufp8sf_scale_to_float
-
-    tensor_bf16 = tensor.to(torch.bfloat16)
-    fp4_packed, sf = fp4_quantize(
-        tensor_bf16,
-        global_scale=global_scale,
-        sf_vec_size=sf_vec_size,
-        is_sf_swizzled_layout=False,
-    )
-
-    sf_uint8 = sf.view(torch.uint8).reshape(-1)
-    dequantized = e2m1_and_ufp8sf_scale_to_float(
-        fp4_packed.cpu(),
-        sf_uint8.cpu(),
-        (1.0 / global_scale).cpu(),
-        sf_vec_size=sf_vec_size,
-        ufp8_type=1,
-        is_sf_swizzled_layout=False,
-    ).to(tensor.device)
-
-    return dequantized.float()
-
-
-def quant_dequant_fp4_per_token_reference(
-    tensor: torch.Tensor,
-    global_scale_inv: torch.Tensor,
-    sf_vec_size: int = 16,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """Reference per-token FP4 quantization for FC2 inputs.
-
-    Returns the dequantized tensor before applying the row scale, plus the
-    returned per-token scale. This mirrors the production MoE path where GEMM2
-    consumes the quantized activation and the finalize kernel applies the row
-    scale with the routing scale.
-    """
-    from flashinfer.quantization import (
-        SfLayout,
-        e2m1_and_ufp8sf_scale_to_float,
-        nvfp4_quantize,
-    )
-
-    tensor_bf16 = tensor.to(torch.bfloat16)
-    fp4_packed, sf, per_token_scale = nvfp4_quantize(
-        tensor_bf16,
-        global_scale_inv,
-        sfLayout=SfLayout.layout_linear,
-        sf_vec_size=sf_vec_size,
-        backend="cuda",
-        per_token_activation=True,
-    )
-
-    dequantized = e2m1_and_ufp8sf_scale_to_float(
-        fp4_packed.cpu(),
-        sf.view(torch.uint8).cpu(),
-        torch.ones(1, dtype=torch.float32),
-        sf_vec_size=sf_vec_size,
-        ufp8_type=1,
-        is_sf_swizzled_layout=False,
-    ).to(tensor.device)
-
-    return dequantized.float(), per_token_scale.to(tensor.device)
-
-
-def compute_reference_moe_fp4(
-    hidden_states: torch.Tensor,
-    gemm1_weights: torch.Tensor,
-    gemm2_weights: torch.Tensor,
-    gemm1_alpha: torch.Tensor,
-    gemm2_alpha: torch.Tensor,
-    token_selected_experts: torch.Tensor,
-    token_final_scales: torch.Tensor,
-    num_tokens: int,
-    num_experts: int,
-    top_k: int,
-    hidden_size: int,
-    intermediate_size: int,
-    fc2_input_scale: torch.Tensor = None,
-    use_per_token_activation: bool = False,
-    num_local_experts: int = None,
-    local_expert_offset: int = 0,
-    activation_type: int = ActivationType.Swiglu.value,
-    swiglu_alpha: float = DEFAULT_SWIGLU_ALPHA,
-    swiglu_beta: float = DEFAULT_SWIGLU_BETA,
-    swiglu_limit: float = DEFAULT_SWIGLU_LIMIT,
-) -> torch.Tensor:
-    """Compute reference MoE output using PyTorch operations on GPU.
-
-    Args:
-        hidden_states: Input hidden states [num_tokens, hidden_size]
-        gemm1_weights: GEMM1 weights [num_local_experts, 2*intermediate_size, hidden_size]
-        gemm2_weights: GEMM2 weights [num_local_experts, hidden_size, intermediate_size]
-        gemm1_alpha: GEMM1 per-expert scalar scales [num_local_experts]
-        gemm2_alpha: GEMM2 per-expert scalar scales [num_local_experts]
-        token_selected_experts: Selected expert IDs (global) [num_tokens, top_k]
-        token_final_scales: Routing weights [num_tokens, top_k]
-        num_tokens: Number of tokens
-        num_experts: Total number of experts (global)
-        top_k: Number of experts per token
-        hidden_size: Hidden dimension
-        intermediate_size: Intermediate dimension
-        fc2_input_scale: Optional scale for FC2 input quantization
-        use_per_token_activation: Use per-token activation.
-        num_local_experts: Number of local experts (for EP). Defaults to num_experts.
-        local_expert_offset: Starting expert ID for this EP rank. Defaults to 0.
-        activation_type: GEMM1 activation type. Use ActivationType.Swiglu for
-            gated SwiGLU/OAI and ActivationType.Relu2 for non-gated ReLU^2.
-        swiglu_alpha: SwiGLU sigmoid multiplier.
-        swiglu_beta: SwiGLU up-projection bias.
-        swiglu_limit: SwiGLU clamp limit.
-
-    Returns:
-        Output tensor [num_tokens, hidden_size]
-    """
-    activation_type, gated = normalize_cute_dsl_moe_activation_type(activation_type)
-
-    if num_local_experts is None:
-        num_local_experts = num_experts
-
-    device = hidden_states.device
-
-    hidden_states = hidden_states.float()
-    gemm1_weights = gemm1_weights.float()
-    gemm2_weights = gemm2_weights.float()
-    gemm1_alpha = gemm1_alpha.float()
-    gemm2_alpha = gemm2_alpha.float()
-
-    output = torch.zeros((num_tokens, hidden_size), dtype=torch.float32, device=device)
-
-    for token_idx in range(num_tokens):
-        token_input = hidden_states[token_idx : token_idx + 1]
-
-        for k in range(top_k):
-            expert_idx = token_selected_experts[token_idx, k].item()
-            scale = token_final_scales[token_idx, k].item()
-
-            # Skip invalid expert IDs
-            if expert_idx < 0 or expert_idx >= num_experts:
-                continue
-
-            # Convert global expert ID to local index for EP
-            local_idx = expert_idx - local_expert_offset
-            if local_idx < 0 or local_idx >= num_local_experts:
-                # This expert is not on this EP rank, skip
-                continue
-
-            w1 = gemm1_weights[local_idx]
-            gemm1_out = gemm1_alpha[local_idx] * (token_input @ w1.T)
-
-            per_token_scale = None
-            if gated:
-                linear = gemm1_out[:, :intermediate_size]
-                gate = gemm1_out[:, intermediate_size:]
-                gate_clamped = gate.clamp(max=swiglu_limit)
-                linear_clamped = linear.clamp(min=-swiglu_limit, max=swiglu_limit)
-                act_out = (
-                    gate_clamped
-                    * torch.sigmoid(swiglu_alpha * gate_clamped)
-                    * (linear_clamped + swiglu_beta)
-                )
-            else:
-                act_out = torch.relu(gemm1_out) ** 2
-
-            if fc2_input_scale is not None:
-                if use_per_token_activation:
-                    act_out, per_token_scale = quant_dequant_fp4_per_token_reference(
-                        act_out, fc2_input_scale, sf_vec_size=16
-                    )
-                else:
-                    act_out = quant_dequant_fp4_reference(
-                        act_out, fc2_input_scale, sf_vec_size=16
-                    )
-
-            w2 = gemm2_weights[local_idx]
-            gemm2_out = act_out @ w2.T
-
-            output_scale = scale * gemm2_alpha[local_idx]
-            if per_token_scale is not None:
-                output_scale = output_scale * per_token_scale[0]
-            output[token_idx] += output_scale * gemm2_out.squeeze(0)
-
-    return output
-
-
-def create_moe_tensors(
-    num_tokens: int,
-    hidden_size: int,
-    intermediate_size: int,
-    num_experts: int,
-    num_local_experts: int,
-    top_k: int,
-    device: str = "cuda",
-    seed: int = 42,
-    gated: bool = True,
-    use_per_token_activation: bool = False,
-):
-    """Create properly quantized MoE tensors for testing."""
-    from flashinfer.fp4_quantization import fp4_quantize
-    from flashinfer.quantization import (
-        SfLayout,
-        e2m1_and_ufp8sf_scale_to_float,
-        nvfp4_quantize,
-    )
-    from flashinfer.quantization.nvfp4_quantization_utils import (
-        current_nvfp4_4over6_config,
-        make_nvfp4_global_scale,
-    )
-    from flashinfer.cute_dsl.utils import convert_sf_to_mma_layout
-
-    torch.manual_seed(seed)
-    sf_vec_size = 16
-
-    # Input
-    x_bf16 = (
-        torch.randn(num_tokens, hidden_size, dtype=torch.bfloat16, device=device) / 10
-    )
-    x_per_token_scale = None
-    if use_per_token_activation:
-        x_global_scale = make_nvfp4_global_scale(
-            x_bf16,
-            per_token_activation=True,
-            nvfp4_4over6_config=current_nvfp4_4over6_config(),
-        )
-        x_quantized, x_sf, x_per_token_scale = nvfp4_quantize(
-            x_bf16,
-            x_global_scale,
-            sfLayout=SfLayout.layout_linear,
-            sf_vec_size=sf_vec_size,
-            backend="cuda",
-            per_token_activation=True,
-        )
-        x_ref = e2m1_and_ufp8sf_scale_to_float(
-            x_quantized.cpu(),
-            x_sf.view(torch.uint8).cpu().reshape(-1),
-            torch.ones(1, dtype=torch.float32),
-            sf_vec_size=sf_vec_size,
-            ufp8_type=1,
-            is_sf_swizzled_layout=False,
-        ).to(device)
-        x_ref = x_ref.float() * x_per_token_scale.unsqueeze(1)
-    else:
-        a1_gs = torch.tensor([1.0], device=device, dtype=torch.float32)
-        x_quantized, x_sf = fp4_quantize(
-            x_bf16,
-            global_scale=a1_gs,
-            sf_vec_size=sf_vec_size,
-            is_sf_swizzled_layout=False,
-        )
-        x_ref = x_bf16.float()
-    x_sf = x_sf.unsqueeze(-1)
-
-    # Routing
-    router_logits = torch.randn(num_tokens, num_experts, device=device)
-    routing_weights = F.softmax(router_logits, dim=1, dtype=torch.float)
-    routing_weights, selected_experts = torch.topk(routing_weights, top_k, dim=-1)
-    routing_weights = routing_weights / routing_weights.sum(dim=-1, keepdim=True)
-    routing_weights = routing_weights.float()
-    selected_experts = selected_experts.to(torch.int32)
-
-    # GEMM1 weights: gated SwiGLU has 2*intermediate rows (interleaved
-    # linear+gate); non-gated ReLU^2 has a single intermediate-row projection.
-    fc1_rows = 2 * intermediate_size if gated else intermediate_size
-    w1_bf16 = (
-        torch.randn(
-            num_local_experts,
-            fc1_rows,
-            hidden_size,
-            dtype=torch.bfloat16,
-            device=device,
-        )
-        / 10
-    )
-
-    w1_gs = torch.tensor([1.0], device=device, dtype=torch.float32)
-    w1_for_quant = (
-        interleave_linear_and_gate(w1_bf16, group_size=64, dim=1) if gated else w1_bf16
-    )
-    w1_flat = w1_for_quant.reshape(num_local_experts * fc1_rows, hidden_size)
-    w1_q_flat, w1_sf_flat = fp4_quantize(
-        w1_flat, global_scale=w1_gs, sf_vec_size=sf_vec_size, is_sf_swizzled_layout=True
-    )
-    w1_q = w1_q_flat.view(num_local_experts, fc1_rows, hidden_size // 2)
-    w1_weight_sf = convert_sf_to_mma_layout(
-        w1_sf_flat,
-        m=fc1_rows,
-        k=hidden_size,
-        num_groups=num_local_experts,
-        sf_vec_size=sf_vec_size,
-    )
-    w1_alpha = torch.linspace(
-        0.75, 1.25, num_local_experts, device=device, dtype=torch.float32
-    )
-
-    # GEMM2 weights
-    w2_bf16 = (
-        torch.randn(
-            num_local_experts,
-            hidden_size,
-            intermediate_size,
-            dtype=torch.bfloat16,
-            device=device,
-        )
-        / 10
-    )
-
-    w2_gs = torch.tensor([1.0], device=device, dtype=torch.float32)
-    w2_flat = w2_bf16.view(num_local_experts * hidden_size, intermediate_size)
-    w2_q_flat, w2_sf_flat = fp4_quantize(
-        w2_flat, global_scale=w2_gs, sf_vec_size=sf_vec_size, is_sf_swizzled_layout=True
-    )
-    w2_q = w2_q_flat.view(num_local_experts, hidden_size, intermediate_size // 2)
-    w2_weight_sf = convert_sf_to_mma_layout(
-        w2_sf_flat,
-        m=hidden_size,
-        k=intermediate_size,
-        num_groups=num_local_experts,
-        sf_vec_size=sf_vec_size,
-    )
-    w2_alpha = torch.linspace(
-        1.25, 0.75, num_local_experts, device=device, dtype=torch.float32
-    )
-
-    fc2_input_scale = torch.tensor([1.0], device=device, dtype=torch.float32)
-
-    return {
-        "x": x_quantized,
-        "x_sf": x_sf,
-        "x_bf16": x_bf16,
-        "x_ref": x_ref,
-        "x_per_token_scale": x_per_token_scale,
-        "token_selected_experts": selected_experts,
-        "token_final_scales": routing_weights,
-        "w1_weight": w1_q,
-        "w1_weight_sf": w1_weight_sf,
-        "w1_weight_bf16": w1_bf16,
-        "w1_alpha": w1_alpha,
-        "fc2_input_scale": fc2_input_scale,
-        "w2_weight": w2_q,
-        "w2_weight_sf": w2_weight_sf,
-        "w2_weight_bf16": w2_bf16,
-        "w2_alpha": w2_alpha,
-    }
-
-
-def check_accuracy(
-    actual: torch.Tensor, expected: torch.Tensor, percent_threshold: float = 0.97
-):
-    """Check numerical accuracy with percentage-based tolerance.
-
-    Tolerances are scaled by output magnitude to account for FP4 quantization
-    noise growing with larger hidden dimensions.
-    """
-    actual = actual.float()
-    expected = expected.float()
-
-    output_scale = max(expected.std().item(), 0.01)
-    atol = max(0.05, 1.5 * output_scale)
-    rtol = 0.5
-
-    abs_diff = torch.abs(actual - expected)
-    rel_diff = abs_diff / (torch.abs(expected) + 1e-8)
-    within_tolerance = (abs_diff < atol) | (rel_diff < rtol)
-    percent_within = within_tolerance.float().mean().item()
-
-    return percent_within >= percent_threshold, percent_within, atol
 
 
 # =============================================================================
